record deleting
fixed wrong JSON format between front and back. Changed accepting format in the back, cleaned up the code wherever needed.

The issue still remains - frontend either displays an incorrect error message when deleting a record and fails to update the user, or there is an underlying issue with worklog updating.

There is also another issue for some reason which was not present in Opera GX browser but is present in the Google Chrome - user fetching issue. Retrieving user from the corresponding page fails to fetch the user, but the userProfile retrieval fetches it correctly.