dynamic profile update
updating profile and all stats when a record is added/deleted from the userProfile page. Fixed a bug which wouldn't reload the user stats and/or update it partially only.

Unified user loading for more consistent code - now each user retrieval updates the local memory and the user profile is loaded from the said memory always.