auth prep, navbar, minor adjustments
preparing frontend for the auth system to be added.
Created login and register pages. Adding a record shouldn't work if the user isn't logged in. Now user should only be able to add their own record. For now, it will get a permission denied, will probably remove the buttons completely later on.
Some minor fixes are still needed to be done, but not that important.