creds validity & availability checks
Added logic to check if username and/or email is valid and available(ie not taken).

email and username are both checked once the user finishes typing in the corresponding field, and again checked once the final request is sent.

Moved a bit of functionality between util files to fit file names more.